Attacks? No, animations.
The tutorial im using for melee combat (https://www.youtube.com/watch?v=rwO3TE1G3ag&ab_channel=MoreBBlakeyyy) uses animations a lot for its combat. which sucks, becuase that is definitely not the stage of the game I'm in, and i got too absorbed into trying to make sprites because it was fun. So uh, now theres animations that may even not be permenant that i put most of my day into. Whoopee.

- Added Idle animation, attack animation (not working)
- Added Animator to player and uncommented animator elements in PlayerController
- Dude I dont know i made a foolish mistake trying to work on sprites today. I am so disappointed in myself