Attack is still an issue, moving on!
I have been wasting a lot of time working on fixing this attack animation so Mr. Buckley told me to move on for now. I added a GUI for the dimensional shards, and I plan to work on making that mechanic functional. Enemy pathfinding comes after, bless my soul.

Continue by continuing to make the Score script from the tutorial website, then make sure it isn't just score and is actually a requirement and make sure that works with the door n stuff